renamed EncounterBuilder components
